feat(landing): mount standalone living-house landing at /
Adds LivingHouseLanding composition and moves the homepage out of the
(app) route group so the landing renders without the sidebar, mobile
header, or fixed top-right app controls. Internal routes under (app)
continue to inherit the app shell.

Preserves on the new root page:

- import "server-only";
- dynamic = "force-dynamic";
- metadata;
- WebSiteSchema + ProfilePageSchema;
- getBaseUrl, getHelsinkiTimeContext, computeStreak;
- Promise.all of fetchLandingPage, fetchAnalytics, fetchThoughts,
  fetchDreams (fetchLandingSummary removed; landing.content no longer
  rendered).

Deletes apps/web/src/app/(app)/page.tsx to resolve the duplicate /
route.
